DNNS.jl Documentation

Overview

Below is a collection of structures and functions which are useful to create (D)eep (N)eural (N)etworks. One of the features of this package is the ability to do Automatic Differentiation. This is accomplished via the structure, AD. The basic arithmatic functions as well as the power, abs, log, exponential, and all of the trigonometric functions from the base package are overloaded by this package to work with AD "numbers".

Additionally, this package adds a notion of a piece-wise linear function via the struct, PWL. This structure, as well as are the structures DLayer and DNN, has been instrumented so that it can also serve as a function.

Structures:

AD: A structure used to implement Automatic Differentiation.

PWL: A structure representing a piece-wise linear function.

DLayer: A structure representing a single layer of a DNN.

DNN: A structure representing a Deep Neural Net which consists of Some number of DLayers.

Functions

L1

Non-linear Activation Functions

sigmoid1

sigmoid2

sigmoid3

softmax

relu

relur

Associated Learning Functions

loss

fit

Functions on PWL structures

smooth

Data Structures

DNNS.AutoDiff.AD — Type

AD{T}

Automatic differentiation structure. Essentially, an implementation of a dual number: the value of an expression together with the derivative of that expression with respect to one chosen variable. AD values are immutable; to change the value or derivative of a stored AD, store a new one (e.g. v[i] = AD(v[i].v, d)).

Fields

v :: T – The value of this structure.

d :: T – The derivative at this value.

Constructors

AD(v) / AD(v; var=true) – A constant (derivative 0), or the variable itself (derivative 1).

AD(v, d) – Value and derivative (promoted to a common type).

AD{T}(v[, d]) – Value and derivative converted to T.

AD{T}(x::AD) – Convert an AD to element type T.

Two ADs are == when both their values and derivatives are ==; isapprox (≈) compares both fields with the usual atol/rtol keywords.

DNNS.PWLF.PWL — Type

PWL{T}

A structure representing a piece-wise linear function on the Real line.

In practice, one uses one of two outer constructors to create a PWL struct.

Type Constraints

T <: Number

The type T must have a total ordering.

Fields

xs :: Vector{T} – The "x" values.

ys :: Vector{T} – The "y" values.

ds :: Vector{T} – The "slopes" of the segments. Including the left most and right most slopes.

n :: Int – The number of "x,y" values.

Input Contract

xs – Must be in strict ascending order.

xs – $|\bf{xs}| \ge 2$.

ys – $|\bf{xs}| = |\bf{ys}|$.

ds – $|\bf{ds}| = |\bf{xs}| + 1$.

Public Constructors

Inputs are promoted to a common floating point type (integer inputs give a PWL{Float64}).

PWL(xs::Vector{T}, y::T, ds::Vector{T})

xs – The x coordinates in ascending order – no duplicates.

y – The value of y corresponding to the first entry in xs.

ds – The slopes of all "x" intervals as well as the "left" slope of the first point and the "right" slope of the last point.

PWL(xs::Vector{T}, ys::Vector{T}, ds::Vector{T})

xs – The x coordinates in ascending order – no duplicates.

ys – The y coordinates corresponding to each x value.

ds – A 2-Vector consisting of the "left" slope of the first point and the "right" slope of the last point.

Examples

Create the same (in behavior) piecewise linear function in two ways:

julia> pw1 = PWL([1.0, 2.0, 3.0], [2.0, 3.0, 4.0], [0.0, 5.0]);

julia> pw2 = PWL([1.0, 2.0, 3.0], 2.0, [0.0, 1.0, 1.0, 5.0]);

julia> pw1(2.5)
3.5

julia> pw2(2.5)
3.5

DNNS.DLayer — Type

DLayer{T<:Number, F}

A structure representing one layer of a neural net.

Type Constraints

T <: Number

The type T must have a total ordering.

F is the type of the activation function (any callable).

Fields

M :: Matrix{AD{T}} – The weight matrix.

b :: Vector{AD{T}} – The bias vector.

op :: F – The activation (threshold) function, applied element-wise.

dims :: Tuple{Int, Int} – The (output, input) dimensions.

Public Constructors

DLayer(Mn::AbstractMatrix{<:Number}, bn::AbstractVector{<:Number}, opn)

Mn – A MxN weight matrix (N inputs, M outputs).

bn – An M dimensional bias vector.

op – The non-linear threshold function.

The element type T is the promotion of the element types of Mn and bn; DLayer{T}(Mn, bn, opn) converts both to T.

DNNS.DNN — Type

DNN{T<:Number}

A structure representing a neural network.

Type Constraints

T <: Number

The type T must have a total ordering.

Fields

layers :: Vector{DLayer{T}} – The neural net layers.

Public Constructors

DNN(ls::AbstractVector{<:DLayer{T}})

ls – A vector of DLayer (all with the same element type T).

DNNS.PWLF.PWL — Method

(PWL{T})(x::Real) where {T<:Number}

Uses the structure PWL as a piece-wise linear function.

Type Constraints

T <: Number

Arguments

x :: Real – An input value (converted to T).

Return

:: T

(PWL{T})(x::AD{T}) where {T<:Number}

Uses the structure PWL as a piece-wise linear function.

Type Constraints

T <: Number

Arguments

x :: AD{T} – An AutoDiff value.

Return

:: AD{T}

DNNS.PWLF.PWL — Method

(PWL{T})(x::Real) where {T<:Number}

Uses the structure PWL as a piece-wise linear function.

Type Constraints

T <: Number

Arguments

x :: Real – An input value (converted to T).

Return

:: T

(PWL{T})(x::AD{T}) where {T<:Number}

Uses the structure PWL as a piece-wise linear function.

Type Constraints

T <: Number

Arguments

x :: AD{T} – An AutoDiff value.

Return

:: AD{T}

DNNS.DLayer — Method

(L::DLayer{T})(x::AbstractVector) where {T <: Number}

If (N,M) = L.dims, then we may treat the structure, DLayer, as a function: ${\cal R}^M \mapsto {\cal R}^N$ .

Takes input x and passes it through the layer.

Type Constraints

T <: Number

Arguments

x :: AbstractVector – An input vector of dimension M.

Return

::Vector{AD{T}} of dimension N.

DNNS.DNN — Method

(dnn::DNN{T})(x::AbstractVector) where {T <: Number}

Let N = DNN.ls[1].dims[2] and M = DNN.ls[end].dims[1], then here we treat the structure DNN as a function: ${\cal R}^N \mapsto {\cal R}^M$ Takes input x and passes it through each of the layers of DNN.

Type Constraints

T <: Number

Arguments

x :: AbstractVector – An input vector of dimension N (numbers, or AD values).

Return

::Vector{AD{T}} of dimension M.

Non-linear Activation Functions

DNNS.UtilFunc.sigmoid1 — Function

sigmoid1(x::AD{T})

Implements an AD version of the standard "exponential" sigmoid function.

Type Constraints

T <: Number

Arguments

x :: AD{T} – The AD input value.

Return

::AD{T} – The output AD value/derivative.

DNNS.UtilFunc.sigmoid2 — Function

sigmoid2(x::AD{T})

Implements an AD version of the standard "tanh" sigmoid function.

Type Constraints

T <: Number

Arguments

x :: AD{T} – The AD input value.

Return

::AD{T} – The output AD value/derivative.

DNNS.UtilFunc.sigmoid3 — Function

sigmoid3(x::AD{T})

Implements an AD version of the standard "arctan" sigmoid function.

Type Constraints

T <: Number

Arguments

x :: AD{T} – The AD input value.

Return

::AD{T} – The output AD value/derivative.

DNNS.UtilFunc.relu — Function

relu(x::AD{T})

Implements an AD version of the standard relu function.

Type Constraints

T <: Number

Arguments

x :: AD{T} – The AD input value.

Return

::AD{T} – The output AD value/derivative.

DNNS.UtilFunc.relur — Function

relur(x::AD{T}; rng=Random.default_rng())

Implements an AD version of a modified version of the relu function. The modification is that while the value of the relur is the same as relu, its derivative is not. The value of the derivative is 0 or 1, however the boundary moves randomly around the natural input boundary of 0 (one of the offsets ±0.25, ±0.1, ±0.025, ±0.01, 0 is drawn from rng).

Type Constraints

T <: Number

Arguments

x :: AD{T} – The AD input value.

Keyword Arguments

rng :: Random.AbstractRNG – The random number generator used for the boundary offset.

Return

::AD{T} – The output AD value/derivative.

DNNS.UtilFunc.softmax — Function

softmax(x::AbstractVector{T} [, τ=1])

Implements the softmax function. Works for plain numbers and, as an AD version, for a vector of AD values.

Type Constraints

T <: Number

Arguments

x :: AbstractVector{T} – The input vector (plain numbers or AD values).

τ :: Real – The "temperature" parameter (τ > 0).

Return

::Vector{T} – The output vector.

Functions

DNNS.loss — Function

loss(dnn, X, Y)

Computes the loss of the neural network given inputs, X, and outputs Y.

Type Constraints

T <: Number

Arguments

dnn :: DNN{T} – A DNN.

X :: AbstractMatrix – The matrix of input values (one sample per column).

Y :: AbstractMatrix – The matrix of output values (one sample per column).

Return

::AD{T} – The (mean squared) loss of the network.

DNNS.fit — Function

fit(dnn, X, Y; N=1000, relerr=1.0e-6, μ=1.0e-3, verbose=false)

Adjusts the parameters of neural network, dnn, in place to get the best fit of the data: X, Y (gradient descent on the loss, with the gradient computed by forward mode automatic differentiation, one parameter at a time). The parameters of the network are all pairs of matrices and biases for each layer in the network.

Type Constraints

T <: Number

Arguments

dnn :: DNN{T} – The network to fit (modified).

X :: AbstractMatrix – The matrix of input values (one sample per column).

Y :: AbstractMatrix – The matrix of output values (one sample per column).

Keyword Arguments

N::Int=1000 – The maximum number of iterations.

relerr::Real=1.0e-6 – Stop (after 20 iterations) once the relative change of the loss is at most relerr.

μ::Real=1.0e-3 – The learning rate.

verbose::Bool=false – If true, print the loss at each iteration and a summary at the end.

Return

A named tuple (loss, iterations, converged): the final loss, the number of iterations used, and whether the relative-change stopping criterion was met.

DNNS.PWLF.smooth — Function

smooth(p::PWL{T}, Δ::T)

Potentially smooths the PWL object by combining adjacent x nodes if the distance between them is less than Δ. A run of consecutive nodes, each closer than Δ to the previous one, is replaced by a single node whose x and y values are the averages of the x and y values of the nodes in the run. (For two close nodes the new node is half way between them.) The function returns a new (potentially) smoothed PWL struct.

Type Constraints

T <: Number

The type T must have a total ordering.

Arguments

p::PWL{T} – PWL object.

Δ::T – The smoothing window.

Returns

::PWL{T} – New, potentially smoothed PWL struct.

Base.merge — Method

merge(p1::PWL{T}, p2::PWL{T})

Merges two PWL objects into one. The new object contains all of the x (node) values. If the two objects have values for the same x (node) – nodes closer than the relative tolerance used by the PWL constructor count as the same – the corresponding x, y values will be the values from p2. If either end node of p1 and p2 coincide, take, for that node, the derivative value from the corresponding node of p2.

Type Constraints

T <: Number

The type T must have a total ordering.

Arguments

p1::PWL{T} -- FirstPWL` object.

p2::PWL{T} -- SecondPWL` object.

Return

::PWL{T} – A combined PWL object.

RecipesBase.plot — Method

plot(p::PWL{T}; <keywords>)

Plots a PWL object.

Arguments

p::PWL{T} – A PWL object

Keyword Arguments

label::Union{String, Nothing} – Label of the graph

lw::Int – Line width.

lc::Symbol – Line color.

`ec::Symbol – Edge color.

`es::Symbol – Edge drawing style.

DNNS.UtilFunc.L1 — Function

L1(v::AbstractVector{T})

The $L_1$ norm (sum of absolute values) of the vector v; works for plain numbers and for AD values.

Return

The sum of the absolute values of the elements of v (of the element type of v).

Index

DNNS.AutoDiff.AD

DNNS.DLayer

DNNS.DLayer

DNNS.DNN

DNNS.DNN

DNNS.PWLF.PWL

DNNS.PWLF.PWL

DNNS.PWLF.PWL

Base.merge

DNNS.PWLF.smooth

DNNS.UtilFunc.L1

DNNS.UtilFunc.relu

DNNS.UtilFunc.relur

DNNS.UtilFunc.sigmoid1

DNNS.UtilFunc.sigmoid2

DNNS.UtilFunc.sigmoid3

DNNS.UtilFunc.softmax

DNNS.fit

DNNS.loss

RecipesBase.plot