DNNS.jl Documentation

Overview

Below is a collection of structures and functions which are useful to create (D)epp (N)eural (N)etworks. One of the features of this package is the ability to do Automatic Differentiation. This is accomplished via the structure, AD. The basic arithematic functions as well as the power, log, exponential, and all of the trigonometric functions from the base package are overloaded by this package to work with AD "numbers".

Additionaly, this package adds a notion of a piece-wise linear function via the struct, PWL. This structure, as well as are the structures DLayer and DNN, has been instrumented so that it can also serve as a function.

Structures:

PWL: A structure representing a piece-wise linear function.

AD: A structure representing the value and its associated derivative. Such a structure is used to compute "Automatic Differentiation".

DLayer: A structure representing a single layer of a DNN.

DNN: A structure representing a Deep Neural Net.

Functions

Non-linear Activation Functions

sigmoid1

sigmoid2

sigmoid3

relu

relur

Associated Learning Functions

loss

fit

Data Structures

DNNS.AD — Type

AD{T}

Automatic differentiation structure.

Fields

v :: T – The value of this structure.

d :: T – The derivative at this value.

DNNS.PWL — Type

PWL{T}

A structure representing a piece-wise linear function.

In practice, one uses one of two outer constructors to create a PWL struct.

Type Constraints

T <: Number

The type T must have a total ordering.

Fields

xs :: Vector{T} – The "x" values.

ys :: Vector{T} – The "y" values.

ds :: Vector{T} – The "slopes" of the segments.

n :: Int – The number of "x,y" values.

Public Constructors

PWL(xs::Vector{T}, y::T, ls::Vector{T})

xs – The x coordinates in ascending order – no duplicates.

y – The value of y corresponding to the first entry in xs.

ls – The slope values before the first and last xs value.

PWL(xs::Vector{T}, nys::Vector{T}, nls::Vector{T})

xs – The x coordinates in ascending order – no duplicates.

ys – The y coordinates corresponding to each x value.

Examples

julia> # Create the same (in behavior) Piecewise linear functions in two ways:
julia> pw1 = PWL([1.0, 2.0, 3.0], [2.0, 3.0, 4.0], [0.0, 5.0])

julia> pw2 = PWL([1.0, 2.0, 3.0], 2.0, [0.0, 1.0, 1.0, 5.0])

julia> pw1(2.5)
3.5

julia> pw2(2.5)
3.5

DNNS.DLayer — Type

DLayer{T<:Number}

A structure representing one layer of a neural net.

Type Constraints

T <: Number

The type T must have a total ordering.

Fields

M :: Matrix{AD{T}} – The "x" values.

b :: Vector{AD{T}} – The "y" values.

op :: Function – The "slopes" of the segments.

dims :: Tuple{Int, Int} – The number of "x,y" values.

Public Constructors

DLayer(Mn::Matrix{T}, bn::Vector{T}, opn)

Mn – A MxNweight matrix.

bn – A Ndimensional bias vector.

op – The non-linear thresholding function.

DNNS.DNN — Type

DNN{T<:Number}

A structure representing one layer of a neural net.

Type Constraints

T <: Number

The type T must have a total ordering.

Fields

layers :: Vector{DLayer{T} – The neural net layers.

Public Constructors

DNN(ls::Vector{DLayer{T}}

ls – A vector of DLayers.

DNNS.PWL — Method

(PWL{T})(x::T) where {T<:Number}

Uses the structure PWL as a piece-wise linear function.

Type Constraints

T <: Number

Arguments

x :: T – An input value.

Return

:: T

(PWL{T})(x::AD{T}) where {T<:Number}

Uses the structure PWL as a piece-wise linear function.

Type Constraints

T <: Number

Arguments

x :: AD{T} – An AutoDiff value.

Return

:: AD{T}

DNNS.DLayer — Method

(L::DLayer{T})(x::AD{T}) where {T <: Number}

If (M,N) = L.dims, then we may treat the structure, DLayer, as a function: ${\cal R}^m \mapsto {\cal R}^n$ .

Takes input x and passes it through the layer.

Type Constraints

T <: Number

Arguments

x :: AD{T} – An input value of dimension N.

Return

::Vector{AD{T}} of dimension M.

DNNS.DNN — Method

(dnn::DNN{T})(x::AbstractVector{T}) where {T <: Number}

Let N = DNN.ls[1].dims[2] and M = DNN.ls[end].dims[1], then here we treatsthe structure DNN as a function: ${\cal R}^N \mapsto {\cal R}^M$ Takes input x and passes it through each of the layers of DNN.

Type Constraints

T <: Number

Arguments

x :: T – An input value of dimension N.

Return

::Vector{AD{T}} of dimension M.

Non-linear Activation Functions

DNNS.sigmoid1 — Function

sigmoid1(x)

Implements an AD version of the standard "exponential" sigmoid function. j

Type Constraints

T <: Number

Arguments

x :: AD{T} – The AD input value.

Return

::AD{T} – The output AD value/derivative.

DNNS.sigmoid2 — Function

sigmoid2(x)

Implements an AD version of the standard "tanh" sigmoid function.

Type Constraints

T <: Number

Arguments

x :: AD{T} – The AD input value.

Return

::AD{T} – The output AD value/derivative.

DNNS.sigmoid3 — Function

sigmoid3(x)

Implements an AD version of the standard "arctan" sigmoid function.

Type Constraints

T <: Number

Arguments

x :: AD{T} – The AD input value.

Return

::AD{T} – The output AD value/derivative.

DNNS.relu — Function

relu(x)

Implements an AD version of the standard relu function.

Type Constraints

T <: Number

Arguments

x :: AD{T} – The AD input value.

Return

::AD{T} – The output AD value/derivative.

DNNS.relur — Function

relur(x)

Implements an AD version of a nodified version of the relu function. The modification is that while the value of the relur is the same as relu, its derivative is not. The value of the derivative is 0 or 1, however the boundary moves randomly around the natural input boundary of 0, its derivative is not. The value of the derivative is 0 or 1, however the boundary moves randomly around the natural input boundary of 0.

Type Constraints

T <: Number

Arguments

x :: AD{T} – The AD input value.

Return

::AD{T} – The output AD value/derivative.

Functions

DNNS.loss — Function

loss(dnn, X, Y)

Computes the loss of the neural network given inputs, X, and outputs Y.

Type Constraints

T <: Number

Arguments

dnn :: DNN{T} – A DNN layer.

X :: Matrix{T} – The matrix of input values.

Y :: Matrix{T} – The matrix of output values.

Return

::AD{T} – The loss of the network

DNNS.fit — Function

fit(dnn, X, Y)

Adjusts the parameters of neural network, dnn, to get the best fit of the data: X, Y. The parameters of the network are all paris of matrices and biases for each layer in the network.

Type Constraints

T <: Number

Arguments

dnn :: DNN{T} – A DNN layer.

X :: Matrix{T} – The matrix of input values.

Y :: Matrix{T} – The matrix of output values.

Return

::nothing

Index

DNNS.AD

DNNS.DLayer

DNNS.DLayer

DNNS.DNN

DNNS.DNN

DNNS.PWL

DNNS.PWL

DNNS.fit

DNNS.loss

DNNS.relu

DNNS.relur

DNNS.sigmoid1

DNNS.sigmoid2

DNNS.sigmoid3