DNNS.jl Documentation

Overview

Below is a collection of structures and functions which are useful to create (D)eep (N)eural (N)etworks. One of the features of this package is the ability to do Automatic Differentiation. This is accomplished via the structure, AD. The basic arithmatic functions as well as the power, abs, log, exponential, and all of the trigonometric functions from the base package are overloaded by this package to work with AD "numbers".

Additionally, this package adds a notion of a piece-wise linear function via the struct, PWL. This structure, as well as are the structures DLayer and DNN, has been instrumented so that it can also serve as a function.

Structures:

AD: A structure used to implement Automatic Differentiation.

PWL: A structure representing a piece-wise linear function.

DLayer: A structure representing a single layer of a DNN.

DNN: A structure representing a Deep Neural Net which consists of Some number of DLayers.

Functions

L1

Non-linear Activation Functions

sigmoid1

sigmoid2

sigmoid3

softmax

relu

relur

Associated Learning Functions

loss

fit

Functions on PWL structures

smooth

Data Structures

DNNS.AutoDiff.AD — Type

AD{T}

Automatic differentiation structure. Essentailly, an implementation of a dual number.

Fields

v :: T – The value of this structure.

d :: T – The derivative at this value.

DNNS.PWLF.PWL — Type

PWL{T}

A structure representing a piece-wise linear function on the Real line.

In practice, one uses one of two outer constructors to create a PWL struct.

Type Constraints

T <: Number

The type T must have a total ordering.

Fields

xs :: Vector{T} – The "x" values.

ys :: Vector{T} – The "y" values.

ds :: Vector{T} – The "slopes" of the segments. Including the left most and right most slopes.

n :: Int – The number of "x,y" values.

Input Contract

xs – Must be in strict ascending order.

xs – $|\bf{xs}| \ge 2$.

ys – $|\bf{xs}| = |\bf{ys}|$.

ds – $|\bf{ds}| = |\bf{xs}| + 1$.

Public Constructors

PWL(xs::Vector{T}, y::T, ds::Vector{T})

xs – The x coordinates in ascending order – no duplicates.

y – The value of y corresponding to the first entry in xs.

ds – The slopes of all "x" intervals as well as the "left" slope of the first point and the "right" slope of the last point.

PWL(xs::Vector{T}, ys::Vector{T}, ds::Vector{T})

xs – The x coordinates in ascending order – no duplicates.

ys – The y coordinates corresponding to each x value.

ds – A 2-Vector consisting of the "left" slope of the first point and the "right" slope of the last point.

Examples

julia> # Create the same (in behavior) Piecewise linear functions in two ways:
julia> pw1 = PWL([1.0, 2.0, 3.0], [2.0, 3.0, 4.0], [0.0, 5.0])

julia> pw2 = PWL([1.0, 2.0, 3.0], 2.0, [0.0, 1.0, 1.0, 5.0])

julia> pw1(2.5)
3.5

julia> pw2(2.5)
3.5

DNNS.DLayer — Type

DLayer{T<:Number}

A structure representing one layer of a neural net.

Type Constraints

T <: Number

The type T must have a total ordering.

Fields

M :: Matrix{AD{T}} – The "x" values.

b :: Vector{AD{T}} – The "y" values.

op :: Function – The "slopes" of the segments.

dims :: Tuple{Int, Int} – The number of "x,y" values.

Public Constructors

DLayer(Mn::Matrix{T}, bn::Vector{T}, opn::Function)

Mn – A MxNweight matrix.

bn – A Ndimensional bias vector.

op – The non-linear threshold function.

DNNS.DNN — Type

DNN{T<:Number}

A structure representing one layer of a neural net.

Type Constraints

T <: Number

The type T must have a total ordering.

Fields

layers :: Vector{DLayer{T} – The neural net layers.

Public Constructors

DNN(ls::Vector{DLayer{T}}

ls – A vector of DLayer.

DNNS.PWLF.PWL — Method

(PWL{T})(x::T) where {T<:Number}

Uses the structure PWL as a piece-wise linear function.

Type Constraints

T <: Number

Arguments

x :: T – An input value.

Return

:: T

(PWL{T})(x::AD{T}) where {T<:Number}

Uses the structure PWL as a piece-wise linear function.

Type Constraints

T <: Number

Arguments

x :: AD{T} – An AutoDiff value.

Return

:: AD{T}

DNNS.DLayer — Method

(L::DLayer{T})(x::AD{T}) where {T <: Number}

If (M,N) = L.dims, then we may treat the structure, DLayer, as a function: ${\cal R}^m \mapsto {\cal R}^n$ .

Takes input x and passes it through the layer.

Type Constraints

T <: Number

Arguments

x :: AD{T} – An input value of dimension N.

Return

::Vector{AD{T}} of dimension M.

DNNS.DNN — Method

(dnn::DNN{T})(x::AbstractVector{T}) where {T <: Number}

Let N = DNN.ls[1].dims[2] and M = DNN.ls[end].dims[1], then here we treat the structure DNN as a function: ${\cal R}^N \mapsto {\cal R}^M$ Takes input x and passes it through each of the layers of DNN.

Type Constraints

T <: Number

Arguments

x :: T – An input value of dimension N.

Return

::Vector{AD{T}} of dimension M.

Non-linear Activation Functions

DNNS.UtilFunc.sigmoid1 — Function

sigmoid1(x::AD{T})

Implements an AD version of the standard "exponential" sigmoid function.

Type Constraints

T <: Number

Arguments

x :: AD{T} – The AD input value.

Return

::AD{T} – The output AD value/derivative.

DNNS.UtilFunc.sigmoid2 — Function

sigmoid2(x::AD{T})

Implements an AD version of the standard "tanh" sigmoid function.

Type Constraints

T <: Number

Arguments

x :: AD{T} – The AD input value.

Return

::AD{T} – The output AD value/derivative.

DNNS.UtilFunc.sigmoid3 — Function

sigmoid3(x::AD{T})

Implements an AD version of the standard "arctan" sigmoid function.

Type Constraints

T <: Number

Arguments

x :: AD{T} – The AD input value.

Return

::AD{T} – The output AD value/derivative.

DNNS.UtilFunc.relu — Function

relu(x::AD{T})

Implements an AD version of the standard relu function.

Type Constraints

T <: Number

Arguments

x :: AD{T} – The AD input value.

Return

::AD{T} – The output AD value/derivative.

DNNS.UtilFunc.relur — Function

relur(x::AD{T})

Implements an AD version of a modified version of the relu function. The modification is that while the value of the relur is the same as relu, its derivative is not. The value of the derivative is 0 or 1, however the boundary moves randomly around the natural input boundary of 0,

T <: Number

Arguments

x :: AD{T} – The AD input value.

Return

::AD{T} – The output AD value/derivative.

DNNS.UtilFunc.softmax — Function

softmax(x::Vector{T} [, τ=one(T)])

Implements the softmax function.

Type Constraints

T <: Number

Arguments

x :: Vector{T} – The AD input vector.

τ :: T – The "temperature" parameter.

Return

::Vector{T} – The output AD vector.

softmax(x::Vector{AD{T}} [, τ=one(T)])

Implements an AD version of the softmax function.

Type Constraints

T <: Number

Arguments

x :: Vector{AD{T}} – The AD input vector.

τ :: T – The "temperature" parameter.

Return

::Vector{AD{T}} – The output AD vector.

Functions

DNNS.loss — Function

loss(dnn, X, Y)

Computes the loss of the neural network given inputs, X, and outputs Y.

Type Constraints

T <: Number

Arguments

dnn :: DNN{T} – A DNN layer.

X :: Matrix{T} – The matrix of input values.

Y :: Matrix{T} – The matrix of output values.

Return

::AD{T} – The loss of the network

DNNS.fit — Function

fit(dnn, X, Y)

Adjusts the parameters of neural network, dnn, to get the best fit of the data: X, Y. The parameters of the network are all paris of matrices and biases for each layer in the network.

Type Constraints

T <: Number

Arguments

dnn :: DNN{T} – A DNN layer.

X :: Matrix{T} – The matrix of input values.

Y :: Matrix{T} – The matrix of output values.

Return

::nothing

DNNS.PWLF.smooth — Function

smooth(p::PWL{T}, Δ::T)

Potentially smooths the PWL object by combining adjacent x nodes if the distance between them is less than Δ. The new x node replaces the other two and is half way between the other nodes. The new y value is the average of the y values of the other two nodes.

Type Constraints

T <: Number

Arguments

p::PWL{T} – PWL object.

Δ::T – The smoothing window.

Returns

::PWL{T} – New, potentially smoothed PWL object.

Index

DNNS.AutoDiff.AD

DNNS.DLayer

DNNS.DLayer

DNNS.DNN

DNNS.DNN

DNNS.PWLF.PWL

DNNS.PWLF.PWL

DNNS.PWLF.smooth

DNNS.UtilFunc.relu

DNNS.UtilFunc.relur

DNNS.UtilFunc.sigmoid1

DNNS.UtilFunc.sigmoid2

DNNS.UtilFunc.sigmoid3

DNNS.UtilFunc.softmax

DNNS.fit

DNNS.loss